implement get seller profile , acceptproduct ,blockproduct ,unblockproduct ,add products to seller schema ,start with get seller products

+++ controller/product.controller.js
@@ -4,6 +4,7 @@ const Product = require("../models/product.model");
 const appError = require("../utils/appError");
 const httpStatusText = require("../utils/utils");
 const productStatus = require("../utils/productStatus");
+const Seller = require("../models/seller.model");
 
 const getAcceptedProducts = asyncWrapper(async (req, res, next) => {
   const { limit, page, text } = req.query;
@@ -105,7 +106,7 @@ const getPendingProducts = asyncWrapper(async (req, res, next) => {
   });
 });
 
-const acceptProductRequest = asyncWrapper(async (req, res, next) => {
+const acceptProduct = asyncWrapper(async (req, res, next) => {
   const productId = req.params.productId;
   if (!productId) {
     const error = appError.create(
@@ -121,7 +122,10 @@ const acceptProductRequest = asyncWrapper(async (req, res, next) => {
     { _id: productId },
     { status: productStatus.ACCEPTED },
     { new: true }
-  );
+  )
+    .populate("productOwner")
+    .populate("category")
+    .populate("subCategory");
 
   if (!updatedProduct) {
     const error = appError.create(
@@ -139,7 +143,7 @@ const acceptProductRequest = asyncWrapper(async (req, res, next) => {
   });
 });
 
-const blockProductRequest = asyncWrapper(async (req, res, next) => {
+const blockProduct = asyncWrapper(async (req, res, next) => {
   const productId = req.params.productId;
   if (!productId) {
     const error = appError.create(
@@ -155,7 +159,10 @@ const blockProductRequest = asyncWrapper(async (req, res, next) => {
     { _id: productId },
     { status: productStatus.BLOCKED },
     { new: true }
-  );
+  )
+    .populate("productOwner")
+    .populate("category")
+    .populate("subCategory");
 
   if (!updatedProduct) {
     const error = appError.create(
@@ -173,6 +180,43 @@ const blockProductRequest = asyncWrapper(async (req, res, next) => {
   });
 });
 
+const unblockProduct = asyncWrapper(async (req, res, next) => {
+  const productId = req.params.productId;
+  if (!productId) {
+    const error = appError.create(
+      "ProductId is required",
+      400,
+      httpStatusText.FAIL
+    );
+    return next(error);
+  }
+
+  // Update product status to accepted
+  const updatedProduct = await Product.findOneAndUpdate(
+    { _id: productId },
+    { status: productStatus.PENDING },
+    { new: true }
+  )
+    .populate("productOwner")
+    .populate("category")
+    .populate("subCategory");
+
+  if (!updatedProduct) {
+    const error = appError.create(
+      "Product not found",
+      404,
+      httpStatusText.FAIL
+    );
+    return next(error);
+  }
+
+  return res.status(200).json({
+    status: httpStatusText.SUCCESS,
+    data: { product: updatedProduct },
+    message: "Product unBlocked successfully",
+  });
+});
+
 const getSellerProducts = asyncWrapper(async (req, res, next) => {
   const { sellerId, limit, page, text } = req.query;
   const skip = (page - 1) * limit;
@@ -185,7 +229,10 @@ const getSellerProducts = asyncWrapper(async (req, res, next) => {
     .limit(limit)
     .skip(skip);
 
-  const allSellerProducts = await Product.find({ productOwner: sellerId });
+  const allSellerProducts = await Product.find({
+    productOwner: sellerId,
+  });
+
   if (!sellerProducts) {
     const error = appError.create(
       "No products to show",
@@ -225,6 +272,24 @@ const addProduct = asyncWrapper(async (req, res, next) => {
   }
 
   await newProduct.save();
+
+  const sellerId = req?.body?.productOwner;
+  const targetSeller = await Seller.findById(sellerId);
+
+  if (!targetSeller) {
+    const error = appError.create(
+      "seller not found",
+      400,
+      httpStatusText.ERROR
+    );
+    return next(error);
+  }
+
+  targetSeller.products.push(newProduct?._id);
+
+  // Save the updated seller document
+  await targetSeller.save();
+  console.log("targetSeller ====>", targetSeller);
   return res.status(201).json({
     status: httpStatusText.SUCCESS,
     data: { product: newProduct },
@@ -330,6 +395,8 @@ const editProduct = asyncWrapper(async (req, res, next) => {
 const deleteProduct = asyncWrapper(async (req, res, next) => {
   const { productId } = req.params;
   const targetProduct = await Product.findById(productId);
+  const sellerId = targetProduct?.productOwner;
+
   if (!targetProduct) {
     const error = appError.create(
       "Product not found",
@@ -338,6 +405,21 @@ const deleteProduct = asyncWrapper(async (req, res, next) => {
     );
     return next(error);
   }
+
+  const targetSeller = await Seller.findById(sellerId);
+
+  if (!targetSeller) {
+    const error = appError.create(
+      "seller not found",
+      400,
+      httpStatusText.ERROR
+    );
+    return next(error);
+  }
+
+  targetSeller.products.pull(productId);
+  await targetSeller.save();
+
   await Product.deleteOne({ _id: productId });
 
   return res.status(200).json({
@@ -355,7 +437,8 @@ module.exports = {
   deleteProduct,
   getSellerProducts,
   getPendingProducts,
-  acceptProductRequest,
-  blockProductRequest,
+  acceptProduct,
+  blockProduct,
+  unblockProduct,
   getBlockedProducts,
 };
